# apps/hoppify/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hoppify API
  version: 0.1.0
paths:
...
  /beer/search:
    get:
      summary: Search Untappd beers
      description: Returns the top 10 Untappd beer catalog matches using full-text and trigram search.
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
          description: Beer, brewery, or OCR text to search for.
      responses:
        "200":
          description: Beer matches were returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeerSearchResponse"
        "400":
          description: Invalid search query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/beer/search:
    get:
      summary: Search Untappd beers
      description: Versioned alias for `/beer/search`.
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
          description: Beer, brewery, or OCR text to search for.
      responses:
        "200":
          description: Beer matches were returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeerSearchResponse"
        "400":
          description: Invalid search query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/captures:
    get:
      summary: List image captures
      description: Returns stored top-level image captures ordered from newest to oldest.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
          description: Maximum number of captures to return.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of captures to skip.
      responses:
        "200":
          description: Captures were listed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureListResponse"
        "400":
          description: Invalid pagination query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Upload image captures
      description: >
        Stores uploaded JPEG files byte-for-byte without recompressing them. PNG and WebP uploads are
        decoded and stored as JPEG for compatibility with downstream detection and recognition.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items:
                    type: string
                    format: binary
              additionalProperties: false
            encoding:
              files:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "201":
          description: Captures were stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapturesResponse"
        "400":
          description: Invalid request or undecodable image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File or request body is too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported image media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Object storage failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/captures/{uuid}/image:
    get:
      summary: Get a capture image
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Capture or crop UUID.
      responses:
        "200":
          description: JPEG image bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid UUID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Capture UUID was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Capture is not a supported image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Object storage failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/detect:
    post:
      summary: Detect objects in a capture image
      description: >
        Runs the configured rough object detector against exactly one image source: a stored capture UUID,
        an S3 URI returned by captures/crops, or a direct multipart file. The result is an
        Ultralytics-style detection pool intended for rough crop creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectRequest"
            examples:
              uuid:
                summary: Detect by capture UUID
                value:
                  uuid: 018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1
              url:
                summary: Detect by S3 URL
                value:
                  url: s3://hoppify/captures/image/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1.jpg
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
              additionalProperties: false
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "200":
          description: Objects were detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectResponse"
        "400":
          description: Invalid request or undecodable image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Capture UUID was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported image media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Object storage failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Detection model is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/crops:
    get:
      summary: List image crops
      description: Returns stored crop captures ordered from newest to oldest.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
          description: Maximum number of crops to return.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of crops to skip.
      responses:
        "200":
          description: Crops were listed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CropListResponse"
        "400":
          description: Invalid pagination query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create image crops from known bounding boxes
      description: >
        Cuts JPEG crop captures from an existing capture image using bounding boxes
        already produced by detection. When crop refinement is enabled, each rough crop is sent to
        the configured OBB model and the straightened crop is stored.
        Repeating the same parent UUID and bounding boxes returns the same crop captures.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CropRequest"
            examples:
              boxes:
                summary: Crop by capture UUID and boxes
                value:
                  uuid: 018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1
                  boxes:
                    - bbox:
                        - 335.08
                        - 700.39
                        - 414.09
                        - 920.65
                      confidence: 0.76799
      responses:
        "201":
          description: Crop captures were stored or already existed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CropResponse"
        "400":
          description: Invalid request, bbox, or undecodable image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Capture UUID was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported image media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Object storage failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/beer-labels/recognitions:
    get:
      summary: List beer label recognition results
      description: Returns stored model recognition results with the crop capture used as input.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
          description: Maximum number of recognitions to return.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of recognitions to skip.
      responses:
        "200":
          description: Beer label recognitions were listed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeerLabelRecognitionListResponse"
        "400":
          description: Invalid pagination query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/beer-labels/identify:
    post:
      summary: Identify a beer label from a stored image
      description: >
        Runs Gemini 2.5 Flash-Lite v3 against exactly one image source: a stored capture/crop UUID or
        an S3 URI returned by captures/crops. Repeating the same known UUID returns the cached result
        and does not call the model again. The model is not given web search or other external tools.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeerLabelIdentifyRequest"
            examples:
              crop:
                summary: Identify by crop UUID
                value:
                  uuid: 3b7eaa56-f7a6-57e1-bff6-552b103bb263
              url:
                summary: Identify by crop S3 URL
                value:
                  url: s3://hoppify/captures/crops/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1/3b7eaa56-f7a6-57e1-bff6-552b103bb263.jpg
      responses:
        "200":
          description: Beer label recognition result was returned from cache or model inference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeerLabelIdentifyResponse"
        "400":
          description: Invalid JSON request, UUID, or S3 URI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Capture UUID was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported image media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Object storage failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Beer label model is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: LLM inference or internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/beer-labels/identify-batch:
    post:
      summary: Identify beer labels from stored crops
      description: >
        Runs beer label recognition for a list of stored crop UUIDs. The server controls Gemini
        concurrency with a worker pool, applies retries for transient inference errors, and returns
        one item per requested UUID. Individual crop failures are returned inside the response item.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeerLabelIdentifyBatchRequest"
            examples:
              crops:
                summary: Identify multiple crop UUIDs
                value:
                  uuids:
                    - 3b7eaa56-f7a6-57e1-bff6-552b103bb263
                    - 3a9245bb-81ef-5ef5-87df-92240c529936
      responses:
        "200":
          description: Batch recognition completed with per-crop success or error items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeerLabelIdentifyBatchResponse"
        "400":
          description: Invalid JSON request or UUID list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal batch execution error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    BeerLabelIdentifyRequest:
      type: object
      description: Exactly one of `uuid`, `url`, or `uri` is required.
      properties:
        uuid:
          type: string
          format: uuid
          description: Capture or crop UUID returned by `POST /api/v1/captures` or `/api/v1/crops`.
          example: 3b7eaa56-f7a6-57e1-bff6-552b103bb263
        url:
          type: string
          description: S3 URI returned by `POST /api/v1/captures` or `/api/v1/crops`.
          example: s3://hoppify/captures/crops/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1/3b7eaa56-f7a6-57e1-bff6-552b103bb263.jpg
        uri:
          type: string
          deprecated: true
          description: Alias for `url`.
          example: s3://hoppify/captures/crops/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1/3b7eaa56-f7a6-57e1-bff6-552b103bb263.jpg
      additionalProperties: false
    BeerLabelIdentifyResponse:
      type: object
      required:
        - model
        - promptVersion
        - cached
        - result
        - createdAt
      properties:
        uuid:
          type: string
          format: uuid
          description: Present when the source image maps to a known capture or crop UUID.
          example: 3b7eaa56-f7a6-57e1-bff6-552b103bb263
        url:
          type: string
          description: Present when the request used an S3 URI source.
          example: s3://hoppify/captures/crops/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1/3b7eaa56-f7a6-57e1-bff6-552b103bb263.jpg
        model:
          type: string
          example: gemini-2.5-flash-lite
        promptVersion:
          type: string
          example: beer-label-v3-gemini-2.5-flash-lite
        cached:
          type: boolean
          description: True when the result was read from Postgres without calling the model.
          example: false
        result:
          $ref: "#/components/schemas/BeerLabelResult"
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    BeerLabelIdentifyBatchRequest:
      type: object
      required:
        - uuids
      properties:
        uuids:
          type: array
          minItems: 1
          maxItems: 300
          items:
            type: string
            format: uuid
          example:
            - 3b7eaa56-f7a6-57e1-bff6-552b103bb263
            - 3a9245bb-81ef-5ef5-87df-92240c529936
      additionalProperties: false
    BeerLabelIdentifyBatchResponse:
      type: object
      required:
        - recognitions
      properties:
        recognitions:
          type: array
          items:
            $ref: "#/components/schemas/BeerLabelIdentifyBatchItem"
      additionalProperties: false
...
    BeerLabelRecognitionListResponse:
      type: object
      required:
        - recognitions
        - limit
        - offset
        - hasMore
      properties:
        recognitions:
          type: array
          items:
            $ref: "#/components/schemas/BeerLabelRecognitionListItem"
        limit:
          type: integer
          example: 30
        offset:
          type: integer
          example: 0
        nextOffset:
          type: integer
          example: 30
        hasMore:
          type: boolean
          example: true
      additionalProperties: false
...
    DetectRequest:
      type: object
      description: Exactly one of `uuid`, `url`, or `uri` is required for JSON requests.
      properties:
        uuid:
          type: string
          format: uuid
          description: Capture UUID returned by `POST /api/v1/captures`.
          example: 018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1
        url:
          type: string
          description: S3 URI returned by `POST /api/v1/captures` or `/api/v1/crops`.
          example: s3://hoppify/captures/image/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1.jpg
        uri:
          type: string
          deprecated: true
          description: Alias for `url`.
          example: s3://hoppify/captures/image/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1.jpg
      additionalProperties: false
    CropRequest:
      type: object
      required:
        - uuid
        - boxes
      properties:
        uuid:
          type: string
          format: uuid
          description: Parent capture UUID returned by `POST /api/v1/captures` or `/api/v1/crops`.
          example: 018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1
        boxes:
          type: array
          minItems: 1
          maxItems: 300
          items:
            $ref: "#/components/schemas/CropBox"
      additionalProperties: false
    CropBox:
      type: object
      required:
        - bbox
      properties:
        bbox:
          type: array
          description: Absolute `[x1, y1, x2, y2]` crop rectangle in parent image pixels.
          minItems: 4
          maxItems: 4
          items:
            type: number
            format: float
          example:
            - 335.08
            - 700.39
            - 414.09
            - 920.65
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Optional detector confidence carried by the caller. It is not persisted.
          example: 0.76799
      additionalProperties: false
    CropResponse:
      type: object
      required:
        - uuid
        - crops
      properties:
        uuid:
          type: string
          format: uuid
          description: Parent capture UUID used as the crop source.
          example: 018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1
        crops:
          type: array
          items:
            $ref: "#/components/schemas/CropCapture"
      additionalProperties: false
    CropCapture:
      type: object
      required:
        - uuid
        - type
        - uri
        - url
      properties:
        uuid:
          type: string
          format: uuid
          example: 3b7eaa56-f7a6-57e1-bff6-552b103bb263
        type:
          type: string
          example: image_crop
        uri:
          type: string
          example: s3://hoppify/captures/crops/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1/3b7eaa56-f7a6-57e1-bff6-552b103bb263.jpg
        url:
          type: string
          example: s3://hoppify/captures/crops/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1/3b7eaa56-f7a6-57e1-bff6-552b103bb263.jpg
    DetectResponse:
      type: object
      required:
        - images
        - metadata
      properties:
        images:
          type: array
          description: List of processed images. Hoppify detect processes one image per request.
          minItems: 1
          maxItems: 1
          items:
            $ref: "#/components/schemas/DetectImageResult"
        metadata:
          $ref: "#/components/schemas/DetectMetadata"
      additionalProperties: false
    DetectImageResult:
      type: object
      required:
        - shape
        - results
        - speed
      properties:
        shape:
          type: array
          description: Original image dimensions in Ultralytics `[height, width]` format.
          minItems: 2
          maxItems: 2
          items:
            type: integer
            minimum: 1
          example:
            - 720
            - 1280
        results:
          type: array
          description: Detections from all configured detector models, ordered by descending confidence after NMS.
          items:
            $ref: "#/components/schemas/Detection"
        speed:
          $ref: "#/components/schemas/DetectSpeed"
      additionalProperties: false
...
    DetectMetadata:
      type: object
      required:
        - imageCount
        - functionTimeCall
      properties:
        uuid:
          type: string
          format: uuid
          description: Present when the source image maps to a capture or crop UUID.
          example: 018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1
        url:
          type: string
          description: S3 URI used as the inference source when available.
          example: s3://hoppify/captures/image/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1.jpg
        imageCount:
          type: integer
          minimum: 1
          example: 1
        functionTimeCall:
          type: number
          format: float
          minimum: 0
          description: Total request processing time in seconds.
          example: 0.018
      additionalProperties: false
...
    CapturesResponse:
      type: object
      required:
        - captures
      properties:
        captures:
          type: array
          items:
            $ref: "#/components/schemas/Capture"
    Capture:
      type: object
      required:
        - uuid
        - type
        - uri
        - url
      properties:
        uuid:
          type: string
          format: uuid
          example: 018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1
        type:
          type: string
          example: image
        uri:
          type: string
          example: s3://hoppify/captures/image/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1.jpg
        url:
          type: string
          example: s3://hoppify/captures/image/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1.jpg
    CaptureListResponse:
      type: object
      required:
        - captures
        - limit
        - offset
        - hasMore
      properties:
        captures:
          type: array
          items:
            $ref: "#/components/schemas/CaptureListItem"
        limit:
          type: integer
          example: 30
        offset:
          type: integer
          example: 0
        nextOffset:
          type: integer
          example: 30
        hasMore:
          type: boolean
          example: true
    CropListResponse:
      type: object
      required:
        - crops
        - limit
        - offset
        - hasMore
      properties:
        crops:
          type: array
          items:
            $ref: "#/components/schemas/CaptureListItem"
        limit:
          type: integer
          example: 30
        offset:
          type: integer
          example: 0
        nextOffset:
          type: integer
          example: 30
        hasMore:
          type: boolean
          example: true
    BeerSearchResponse:
      type: object
      required:
        - query
        - results
      properties:
        query:
          type: string
          example: Punk IPA BrewDog
        results:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/BeerSearchResult"
    BeerSearchResult:
      type: object
      required:
        - untappdId
        - url
        - untappdSlug
        - searchText
        - lastModifiedAt
        - textRank
        - fuzzyRank
      properties:
        untappdId:
          type: integer
          format: int64
          example: 123
        url:
          type: string
          format: uri
          example: https://untappd.com/b/brewdog-punk-ipa/123
        untappdSlug:
          type: string
          example: brewdog-punk-ipa
        breweryPrefix:
          type: string
          example: brewdog
        searchText:
          type: string
          example: brewdog punk ipa
        lastModifiedAt:
          type: string
          format: date-time
        textRank:
          type: number
          format: double
          example: 0.42
        fuzzyRank:
          type: number
          format: double
          example: 0.74
    CaptureListItem:
      type: object
      required:
        - uuid
        - type
        - uri
        - imageUrl
        - createdAt
        - sizeBytes
      properties:
        uuid:
          type: string
          format: uuid
          example: 018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1
        type:
          type: string
          example: image
        uri:
          type: string
          example: s3://hoppify/captures/image/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1.jpg
        imageUrl:
          type: string
          example: /api/v1/captures/018f7b8e-4d96-7b42-9f64-09e5d3a8e7c1/image
        createdAt:
          type: string
          format: date-time
        sizeBytes:
          type: integer
          format: int64
          example: 245701
        width:
          type: integer
          example: 4032
        height:
          type: integer
          example: 3024
        originalFilename:
          type: string
          example: shelf.png
        originalSizeBytes:
          type: integer
          format: int64
          example: 3750462
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - payload_too_large
                - unsupported_media_type
                - storage_error
                - not_found
                - model_unavailable
                - inference_error
                - internal_error
            message:
              type: string